Agent/weather_agent/requirements.txt
# Core application libraries
langchain>=0.1.0,<0.2.0
langchain-openai>=0.1.0,<0.2.0
deepseek==1.0.0
httpx>=0.27.0,<0.28.0
pyyaml>=6.0,<7.0
schedule>=1.2.0,<1.3.0